Split GroupManagementPage components into files
This commit simply splits major or connected components used by
GroupManagementPage into their own files. No refactorings were made.
