+++ src/components/GroupManagementPage/index.js
@@ -1,0 +1,2 @@
+import GroupManagementPage from './GroupManagementPage'
+export default GroupManagementPage
